% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_database_evidence}
\alias{plot_database_evidence}
\title{Database-evidence composition of primary serum-derived proteins}
\usage{
plot_database_evidence(primary_df)
}
\arguments{
\item{primary_df}{primary serum-derived proteins (merged across groups)}
}
\value{
ggplot object
}
\description{
Database-evidence composition of primary serum-derived proteins
}
